requirements.txt
fastapi>=0.68.0,<1.0.0
supabase>=2.0.0,<3.0.0
python-multipart
uvicorn
pydantic>=1.8.0,<2.0.0
python-dotenv
httpx
gunicorn

# 代码质量工具
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# 测试和覆盖率
pytest-cov>=4.0.0
bandit>=1.7.0  # 安全检查

# 开发工具
ruff>=0.1.0  # 现代化的更快的linter
